code/test/requirements.txt
# File: tests/requirements.txt
pytest==7.4.0
pytest-html==4.0.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-timeout==2.1.0
behave==1.2.6
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1
jsonschema==4.19.0
beautifulsoup4==4.12.2
lxml==4.9.3
regex==2023.8.8